Moving sql patch statement generation APIs to the sqlfmt package and cleaning up import cycles
